Rename EntitiesRegistry related files
