+++ src/app/core/entities/entities-registry.ts
